+++ formio/models/formio_builder.py
@@ -516,6 +516,10 @@ class Builder(models.Model):
         # only set language if exist in i18n translations
         if options['i18n'].get(language):
             options['language'] = language
+        elif self.language_en_enable:
+            lang_en = self.env.ref('base.lang_en')
+            options['language'] = Lang._formio_ietf_code(lang_en.code)
+
         return options
 
     def _get_form_js_locales(self):